% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_utils.R
\name{add_evolution_trace}
\alias{add_evolution_trace}
\title{add_evolution_trace}
\usage{
add_evolution_trace(
  som_obj,
  path,
  point_col = "red",
  line_thickness = 3,
  point_size = 2
)
}
\arguments{
\item{som_obj}{should be a \code{kohonen} object}

\item{path}{the path of neuron want to show on the plot}

\item{point_col}{color of center of points}

\item{line_thickness}{the thickness of connecting lines same as \code{lwd} argument in \code{lines}}

\item{point_size}{the size of points same as  arguments \code{cex} argument in \code{points}}
}
\description{
add_evolution_trace
}
\author{
Hamid Davoukhani \email{h.davoudkhani@gmail.com}
}
