% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_utils.R
\name{add_neuron_population}
\alias{add_neuron_population}
\title{add populaiton of each neuron to the som plots}
\usage{
add_neuron_population(som_obj, text = FALSE, shape = c("hex", "round"))
}
\arguments{
\item{som_obj}{should be a \code{kohonen} object}

\item{text}{logical , if TRUE print the population of each neuron as number}

\item{shape}{a chosen shape size of which show the popluation of each neurons}
}
\value{

}
\description{
add populaiton of each neuron to the som plots
}
\examples{
load("data/pdb_and_tarj_object.RData")

## neuron population

som_trj <- som(trj1 , grid =  somgrid( 6 , 6, "hexagonal"))
som_hc <- cutree(hclust(dist(som_trj$codes[[1]] ), method = "ward.D") , 5)
plot_clustered_map (som_trj,
                    cluster = som_hc,
                    shape = "straight" )
add_neuron_population(som_trj , shape = "round" )

}
\author{
Hamid Davoukhani \email{h.davoudkhani@gmail.com}
}
