% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average.neur.property.r
\name{average.neur.property}
\alias{average.neur.property}
\title{Function to plot the silhouette score for the clustering of SOM neurons}
\usage{
average.neur.property(SOM, P)
}
\arguments{
\item{SOM:}{the SOM object to cluster}

\item{P:}{the property for each frame of the simulation}
}
\value{
D the set of distances used to train the SOM is computed for all the frames.
}
\description{
Function to plot the silhouette score for the clustering of SOM neurons
}
\author{
Stefano Motta\email{stefano.motta@unimib.it}
}
