% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.dist.mat.R
\name{calc.dist.mat}
\alias{calc.dist.mat}
\title{Calculation of Distance matrix}
\usage{
calc.dist.mat(coord)
}
\arguments{
\item{coord}{matrix of N atomic coordinates (N rows, 3 columns)}
}
\value{
mat
}
\description{
Compute the pairwise distance matrix of a given set of coordinates
}
\author{
Stefano Motta\email{stefano.motta@unimib.it}
}
