% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.dists.R
\name{calc.dists}
\alias{calc.dists}
\title{Calculation of Distances}
\usage{
calc.dists(coord, mol.1_id = FALSE, mol.2_id = FALSE, sele = FALSE)
}
\arguments{
\item{coord}{matrix of N atomic coordinates (N rows, 3 columns)}

\item{mol.1_id}{vector containing the index of the first molecule for intermolecular distances only}

\item{mol.2_id}{vector containing the index of the second molecule for intermolecular distances only}

\item{sele}{contains the selection of distances coming from the native_contacts function}
}
\value{
A matrix contaning the selected distances for a frame
}
\description{
Compute the pairwise distance matrix of a given set of coordinates, and only retain to some selected distances
}
\author{
Stefano Motta\email{stefano.motta@unimib.it}
}
