% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_dist_mat.r
\name{Calc_Dist_Mat}
\alias{Calc_Dist_Mat}
\title{Calculation of Distance matrix}
\usage{
Calc_Dist_Mat(coord)
}
\arguments{
\item{coord}{matrix of N atomic coordinates (N rows, 3 columns)}
}
\value{
MAT
}
\description{
Calculation of Distance matrix
}
\author{
Stefano Motta\email{stefano.motta@unimib.it}
}
