% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_distances.r
\name{calc_distances}
\alias{calc_distances}
\title{Function to compute distances to be used to train the SOM}
\usage{
calc_distances(trj, stride = 1, MOL2 = FALSE, sele = FALSE, atoms = NULL)
}
\arguments{
\item{trj}{contains the trajectory coordinates (array with three dimensions obtained by rioxdr)}

\item{stride}{can stride frames (used to stride the trajectory)}

\item{MOL2}{contains the atom indexes of the second molecule in case only intermolecular distances should be computed}

\item{sele}{contains the selection of distances coming from the native_contacts function}
}
\value{
D the set of distances used to train the SOM is computed for all the frames.
}
\description{
Function to compute distances to be used to train the SOM
}
\author{
Stefano Motta\email{stefano.motta@unimib.it}
}
