% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat.trj.r
\name{cat.trj}
\alias{cat.trj}
\title{Concatenate simulations}
\usage{
cat.trj(trj1, ...)
}
\arguments{
\item{trj1}{the first trj file}

\item{...}{additional trj files}
}
\value{
trj with the simulations concatenated
}
\description{
Function to concatenate two simulations.
}
\examples{
# Read the three simulations
trj1 <- read.trj(trjfile="rep_001.xtc", topfile="ref.pdb")
trj2 <- read.trj(trjfile="rep_002.xtc", topfile="ref.pdb")
trj3 <- read.trj(trjfile="rep_003.xtc", topfile="ref.pdb")
# Concatenate the simulations
  trj <- cat.trj(trj1, trj2, trj3)

}
\author{
Stefano Motta \email{stefano.motta@unimib.it}
}
