% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_trj.r
\name{cat_trj}
\alias{cat_trj}
\title{Function to concatenate two simulations.}
\usage{
cat_trj(trj1, ...)
}
\arguments{
\item{trj1}{the first trj file}
}
\value{
trj with the simulations concatenated
}
\description{
Function to concatenate two simulations.
}
\author{
Stefano Motta\email{stefano.motta@unimib.it}
}
