% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.pathways.r
\name{cluster.pathways}
\alias{cluster.pathways}
\title{Clustering of Pathways}
\usage{
cluster.pathways(
  SOM,
  start,
  end,
  time.dep = "independent",
  method = "complete"
)
}
\arguments{
\item{SOM}{a kohonen SOM object.}

\item{start}{the vector specifying the starting frame of each replicas}

\item{end}{the vector specifying the ending frame of each replicas}

\item{time.dep}{choose whether to use time "dependent" or "independent" clustering of pathways}

\item{method}{the method to be passed to hclust for the clustering}
}
\value{
representatives a vector of frames representatives of each neuron
}
\description{
Cluster pathways according to a time dependent or independent scheme
}
\examples{
# Cluster Pathways using the time dependent algorithm
cluster.pathways(SOM, start=trj$start, end=trj$end, time.dep="dependent")

#Cluster Pathways using the time independent algorithm
cluster.pathways(SOM, start=trj$start, end=trj$end, time.dep="dependent")

}
\author{
Stefano Motta\email{stefano.motta@unimib.it}
}
