% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.representatives.r
\name{cluster.representatives}
\alias{cluster.representatives}
\title{Cluster Representatives}
\usage{
cluster.representatives(SOM, clusters)
}
\arguments{
\item{SOM}{a kohonen SOM object.}

\item{clusters}{a vector of clusters assignment for each neuron, as returned for example by hclust.}
}
\value{
representatives a vector of frames representatives of each neuron
}
\description{
Compute the cluster representatives
}
\examples{
# Divide the SOM in the selected number of clusters
#Get representative frames for each cluster
CL_repres <- cluster.representatives(SOM, SOM.hc)
}
\author{
Stefano Motta \email{stefano.motta@unimib.it}
}
