% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_neurons.R
\name{cluster_neuron}
\alias{cluster_neuron}
\title{Title}
\usage{
cluster_neuron(
  som_obj,
  min_cluster = 6,
  max_cluster = 15,
  kohonen_object = NULL
)
}
\arguments{
\item{kohonen_object}{}
}
\value{

}
\description{
Title
}
