% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network.functions.r
\name{comp.trans.mat}
\alias{comp.trans.mat}
\title{Compute transition matrix}
\usage{
comp.trans.mat(classif, start)
}
\arguments{
\item{classif}{a vector of neuron assigment (usually passed by SOM$unit.classif}

\item{start}{a vector containing the start frames of each replica (usually contained in trj$start if replicas were merged with cat_trj)}
}
\value{
trans
}
\description{
Compute the transition matrix starting from a vector of subsequent classifications
}
\author{
Stefano Motta \email{stefano.motta@unimib.it}
}
