% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_mat_frame.R
\name{dist_mat_frame}
\alias{dist_mat_frame}
\title{Title :Calculating the distances between pairs of chosen set  atoms for each frame}
\usage{
dist_mat_frame(pdb, inds = NULL, grp = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{pdb}{: a pdb structure object as returned by read.pdb or a numeric vector of ‘xyz’ coordinates.}

\item{inds}{: atom and xyz coordinate indices obtained from atom.select that selects the elements of pdb upon which the calculation should be based.}

\item{grp}{: logical, if TRUE atomic distances will be grouped according to their residue membership. See ‘grpby’}

\item{...}{}
}
\value{

}
\description{
Title :Calculating the distances between pairs of chosen set  atoms for each frame
}
