% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dRMSD.R
\name{dRMSD}
\alias{dRMSD}
\title{distance root-mean-square deviation}
\usage{
dRMSD(VEC1, VEC2, N_atm)
}
\arguments{
\item{VEC1}{the reference vector}

\item{VEC2}{the second Conformation vector}

\item{N_atm}{the number of atom}
}
\value{

}
\description{
distance root-mean-square deviation
}
\examples{
not defined yet
}
\author{
Hamid Davoukhani \email{h.davoudkhani@gmail.com}

To calculate the root-mean-square deviation between
the atoms in two configurations

\deqn{dRMSD = \sqrt{ 1/N(N-1) * \sum{(X_i - X_j)^2L}

    where i != j}}
X_i coorodinate of Atom i

X_j coorodinate of Atom j

N number of atoms
}
