% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_atom.R
\name{filter_atoms}
\alias{filter_atoms}
\title{Filter pdb file by particular atoms}
\usage{
filter_atoms(pdb_file, atom = "CA", output_type = "pdb")
}
\arguments{
\item{pdb_file}{a \code{bio3d} pdb object}

\item{atom}{the atom which pdb file should be filtered}

\item{output_type}{the output of the function  can be either pdb object which
only contain attributes of the atoms, or a trajectory file contain the atom coordinates}
}
\value{

}
\description{
the function accept a pdb file and filter a given atom, and return either a pdb file or only coordinates

@author Hamid Davoukhani \email{h.davoudkhani@gmail.com}
}
\examples{
pdb1 <- read_pdb("../SOMMD/data/ref_1782.pdb")
trj1 <- read_ncdf("../SOMMD/data/TEST_001.nc")
merged_pdb <-  merge_pdb_trj(pdb1 , trj1)
## pdb outputs
cb_pdb <-  fliter(merged_pdb , atom = "CB", output_type = "pdb")

## coordinates output
cb_coordinates <-  fliter(merged_pdb , atom = "CB", output_type = "xyz")
}
