% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_dist_frame.R
\name{fit_dist_frame}
\alias{fit_dist_frame}
\title{Title :REQ3 : Pre-aligns the trajectory,  using output of requirement (REQ-2) as input}
\usage{
fit_dist_frame(
  fixed,
  mobile,
  fixed.inds = NULL,
  mobile.inds = NULL,
  verbose = FALSE,
  prefix = "",
  pdbext = "",
  outpath = "fitlsq",
  full.pdbs = FALSE,
  ncore = 1,
  nseg.scale = 1,
  ...
)
}
\arguments{
\item{fixed}{:fixed numeric vector of xyz coordinates.}

\item{mobile}{:mobile numeric vector, numeric matrix, or an object with an xyz component containing one or more coordinate sets.}

\item{mobile.inds}{:mobile.inds a vector of indices that selects the elements of mobile upon which fitting should
be based.}

\item{verbose}{:logical, if TRUE more details are printed.}

\item{prefix}{:prefix to mobile$id to locate “full” input PDB files. Only required if full.pdbs
is TRUE.}

\item{pdbext}{:the file name extension of the input PDB files}

\item{outpath}{:character string specifing the output directory when full.pdbs is TRUE}

\item{full.pdbs}{:full.pdbs logical, if TRUE “full” coordinate files (i.e. all atoms) are written to the location
specified by outpath.}

\item{ncore}{:number of CPU cores used to do the calculation. ncore>1 requires package
‘parallel’ installed.}

\item{nseg.scale}{:split input data into specified number of segments prior to running multiple core
calculation.}

\item{fixed.inds:fixed.inds}{a vector of indices that selects the elements of fixed upon which fitting should
be based.}

\item{...other}{parameters for read.pdb.}
}
\description{
Coordinate superposition with the Kabsch algorithm.
The function fit.xyz is a wrapper for the function rot.lsq, which performs the actual coordinate
superposition. The function rot.lsq is an implementation of the Kabsch algorithm (Kabsch, 1978)
and evaluates the optimal rotation matrix to minimize the RMSD between two structures.
}
\references{
Grant, B.J. et al. (2006) Bioinformatics 22, 2695–2696.

:https://www.rdocumentation.org/packages/bio3d/versions/2.4-2/topics/pdbaln
https://rdrr.io/cran/bio3d/man/pdbaln.html
https://search.r-project.org/CRAN/refmans/bio3d/html/pdbaln.html
}
\seealso{
\code{\link[SOMMD]{fit.xyz}}
}
\author{
shaziya Ismail Mulla \email{shaziyaismail.m@gmail.com}
}
