% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gro2pdb.R
\name{gro2pdb}
\alias{gro2pdb}
\title{Convert a gro object into a pdb obtect}
\usage{
gro2pdb(gro)
}
\arguments{
\item{gro}{contains the gro object to convert}
}
\value{
Returns an object with class "pdb"
}
\description{
Convert a gro object into a pdb obtect
}
\examples{
pdb <- gro2pdb(gro)
}
\author{
Stefano Motta \email{stefano.motta@unimib.it}
}
