% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_utils.R
\name{highlight_a_cluster}
\alias{highlight_a_cluster}
\title{Highlight a cluster and its attributes}
\usage{
highlight_a_cluster(
  som_obj,
  clusters,
  cluster_number,
  label = NULL,
  property_value = NULL,
  property_color = NULL,
  label_color = "black",
  col = "red",
  lwd = 4,
  ...
)
}
\arguments{
\item{som_obj}{should be a \code{kohonen} object}

\item{clusters}{a vector of cluster neuron}

\item{cluster_number}{the cluster to be highlighted}

\item{label}{label print at the center of cluster}

\item{property_value}{a different value print at center of different neuron of cluster}

\item{property_color}{color property note, if \code{NULL} use \code{label_color}}

\item{label_color}{color of label note, default is \code{black}}

\item{col}{the color draw a boundary around the cluster}

\item{lwd}{thickness of both notes}

\item{...}{}
}
\value{

}
\description{
Highlight a clusters and add note at center of cluster and a property value in a different center of clusters
}
\examples{
 load("data/pdb_and_tarj_object.RData")
set.seed(100)
som_trj <- SOMMD::som(trj1 , grid =  somgrid( 6 , 6, "hexagonal"))
dummy_property_nuerons <-  rnorm(36 , mean = 8 , sd  = 4)
som_hc <- cutree(hclust(dist(som_trj$codes[[1]] ), method = "ward.D") , 5)


plot_property(som_trj , property = dummy_property_nuerons ,shape = "straight",
              palette_name =  colorRampPalette(c("blue", "white", "red")))
highlight_a_cluster(som_trj,
                    som_hc ,
                    cluster_number = 3,
                    label = "cluster 3" ,
                    property_value = "66",
                    property_color = NULL,
                    label_color = "darkgreen",
                    col = "green",
                    lwd = 2,
                    cex = 2)
}
