% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.functions.R
\name{map.color}
\alias{map.color}
\title{Map the property vector to colours}
\usage{
map.color(x, pal, limits = NULL, na.col = "grey")
}
\arguments{
\item{x}{a numeric vector}

\item{pal}{a color palette}

\item{limits}{the values of the extremes for the colorscale}

\item{na.col}{the color that will be assigned to the na.values of the vector}
}
\value{
A vector with colors proportional to the values of x
}
\description{
Function map a numeric vector of a property to a vector of colors for that property according to that property value.
}
