% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.data.r
\name{map.data}
\alias{map.data}
\title{Map data to SOM}
\usage{
map.data(SOM, X, add = FALSE)
}
\arguments{
\item{SOM}{a trained SOM}

\item{X}{a data set with the same number of features of the dataset used to train the SOM}

\item{add}{whether to append the new data to the ones used to train the SOM}
}
\value{
The SOM with the new data mapped
}
\description{
Assign new data to a trained SOM
}
\examples{
# Compute distances on new simulations (the same used for SOM training)
DIST2 <- calc_distances(trj2, mol.2=FALSE, sele=sele_dists, atoms=sele_atoms)
# Map new data on the existing SOM
SOM_new <- map.data(SOM=SOM, X=DIST2)
}
\author{
Stefano Motta \email{stefano.motta@unimib.it}
}
