% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network.functions.r
\name{map2color}
\alias{map2color}
\title{Function map a numeric vector of a property to a vector of colors for that property according to that property value.}
\usage{
map2color(x, pal, limits = NULL, na.col = "grey")
}
\arguments{
\item{x}{a numeric vector}

\item{pal}{a color palette}

\item{limits}{the values of the extremes for the colorscale}

\item{na.col}{the color that will be assigned to the na.values of the vector}
}
\value{
COL a vector with the same length of x, with colors proportional to the values of x
}
\description{
Function map a numeric vector of a property to a vector of colors for that property according to that property value.
}
