% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network.functions.r
\name{Matrix2Graph}
\alias{Matrix2Graph}
\title{Convert transition matrix to an igraph object}
\usage{
Matrix2Graph(trans, SOM, SOM.hc, COL.SCALE, diag = FALSE)
}
\arguments{
\item{trans}{a transition matrix (usually obtained from comp.trans.mat)}

\item{SOM}{a kohonen object that form the network}

\item{SOM.hc}{a vector of cluster assignment for SOM neurons}

\item{col.set}{a vector of colors used for the SOM clusters}
}
\value{
The network as igraph object, with the SOM properties
}
\description{
Function to convert a transition matrix to an igraph object
}
\author{
Stefano Motta \email{stefano.motta@unimib.it}
}
