% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network.functions.r
\name{Matrix2Network}
\alias{Matrix2Network}
\title{Convert transition matrix to Network}
\usage{
Matrix2Network(trans)
}
\arguments{
\item{trans}{a transition matrix (usually obtained from comp.trans.mat)}
}
\value{
The network in the form of a dataframe
}
\description{
Function to convert a transition matrix to a network
}
\author{
Stefano Motta \email{stefano.motta@unimib.it}
}
