% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network.functions.r
\name{Network2Graph}
\alias{Network2Graph}
\title{convert a network to igraph object}
\usage{
Network2Graph(d, SOM, SOM.hc, col.set)
}
\arguments{
\item{d}{a network in the form of a dataframe}

\item{SOM}{a kohonen object that form the network}

\item{SOM.hc}{a vector of cluster assignment for SOM neurons}

\item{col.set}{a vector of colors used for the SOM clusters}
}
\value{
The network as igraph object, with the SOM properties
}
\description{
Function to convert a network dataframe to an igraph object
}
\author{
Stefano Motta \email{stefano.motta@unimib.it}
}
