% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neur.population.r
\name{neur.population}
\alias{neur.population}
\title{Get Neuron Population}
\usage{
neur.population(SOM)
}
\arguments{
\item{SOM}{the SOM object}
}
\description{
Function to compute the per-neuron population
}
\examples{
neur.population(SOM) 
}
\author{
Stefano Motta \email{stefano.motta@unimib.it}
}
