% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neur.representatives.r
\name{neur.representatives}
\alias{neur.representatives}
\title{Neuron representative}
\usage{
neur.representatives(SOM)
}
\arguments{
\item{SOM}{a kohonen SOM object.}
}
\value{
a vector of frames representatives of each neuron
}
\description{
Compute the representative frame of each neuron (the closest to the neuron codebook)
}
\examples{
NEUR_repres <- neur.representatives(SOM)
}
\author{
Stefano Motta \email{stefano.motta@unimib.it}
}
