% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_matrix.R
\name{normalize_matrix}
\alias{normalize_matrix}
\title{Normalizer the Distance or similar matrices}
\usage{
normalize_matrix(mat, option = "iso")
}
\arguments{
\item{mat}{a given  numerical matrix}

\item{option}{the method of normalization iso or dimension-wise}
}
\value{
a normalized matrices with values between 0 and 1
}
\description{
Normalizer the Distance or similar matrices
}
\examples{
not provided here yet
}
