% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.silhouette.profile.r
\name{plot.silhouette.profile}
\alias{plot.silhouette.profile}
\title{Function to plot the silhouette profile for the Nclus cluster of the SOM neurons}
\usage{
\method{plot}{silhouette.profile}(SOM, Nclus, dist_clust = "euclidean", clust_method = "complete")
}
\arguments{
\item{SOM}{the SOM object to cluster}

\item{Nclus}{the cluster number on which the silhouette profile will be computed}

\item{dist_clust}{the metric for the distance calculation}

\item{clust_method}{the method for the clustering (passed to the hclust function}
}
\value{
D the set of distances used to train the SOM is computed for all the frames.
}
\description{
Function to plot the silhouette profile for the Nclus cluster of the SOM neurons
}
\author{
Stefano Motta\email{stefano.motta@unimib.it}
}
