% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clustered_map.R
\name{plot_clustered_map}
\alias{plot_clustered_map}
\title{Plot clusterd SOM map}
\usage{
plot_clustered_map(
  som_obj,
  cluster = NULL,
  cluster_method = c("hier", "kmeans"),
  clustering_parameter = c(5, "ward.D"),
  bgcols = NULL,
  shape = c("straight", "round")
)
}
\arguments{
\item{som_obj}{should be a \code{kohonen} object}

\item{cluster}{can be a vector showing the number of cluster of each neurons or \code{NULL}}

\item{cluster_method}{if \code{cluster} is \code{NULL} , either \code{"hier"} as hierarchical methods or
\code{"kmeans"} can be chosen to do a basic clustering based on \code{clustering_parameter}}

\item{clustering_parameter}{is a list of to basic clustering hyper-parameter, first number of clusters
and second method of clustering in \code{hier} or algorithm in \verb{kmeans methods}}

\item{bgcols}{hues distinguish clusters can be introduced here}

\item{shape}{shape of neurons , \code{straight} plots neurons as hexagons  and \code{round} as cricles}
}
\description{
Plot clusterd SOM map
}
\author{
Hamid Davoukhani \email{h.davoudkhani@gmail.com}
}
