% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_property.R
\name{plot_property}
\alias{plot_property}
\title{Plot a heatmap based on a given property}
\usage{
plot_property(
  som_obj,
  property,
  FUN = mean,
  main = NULL,
  heatkey = TRUE,
  palette_name = NULL,
  ...
)
}
\arguments{
\item{som_obj}{a \code{kohonen} object}

\item{property}{a vector of property for either each neurons or each observation}

\item{FUN}{when length of property vector is equal to number of observation,
this given argument use as a function to aggregate values for each neurons}

\item{main}{the title of the plot}

\item{heatkey}{logical, when is true, plot the spectrum of the property}

\item{palette_name}{palette color used to hue neurons}

\item{...}{other graphical arguments can be passed to plot funtion}
}
\description{
Plot a heatmap based on a given property
}
\seealso{
\link{plot_clustered_map} ,  \link{\code{add_evolution_trace}} , \link{\code{add_neuron_population}}
}
\author{
Hamid Davoukhani \email{h.davoudkhani@gmail.com}
}
