% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreAlignTrajectory.R
\name{Pre_algn_pdbaln}
\alias{Pre_algn_pdbaln}
\title{Pre Align the Trajectory}
\usage{
Pre_algn_pdbaln(
  files,
  fit = FALSE,
  pqr = FALSE,
  ncore = 1,
  nseg.scale = 1,
  progress = NULL,
  ...
)
}
\arguments{
\item{files}{a character vector of PDB file names. Alternatively, a list of pdb objects can be provided.}

\item{fit}{desc}

\item{pqr}{desc}

\item{ncore}{desc}

\item{nseg.scale}{desc}

\item{progress}{desc}

\item{...}{}
}
\value{

}
\description{
This is a wrap around \link[bio3d]{pdbaln}
This Creates multiple sequences alignments from a list of PDB files returning aligned sequence and structure records
This is written for re-aligns the trajectory,  using output of requirement (REQ-32) as input ,
Returns a list of class "pdbs" with the five  components: xyz, resno, b, chain, id, ali, and call. The coordinates of the aligned trajectory.
the definition of argument duplicated from \link[bio3d]{pdbaln}:

@author Shaziya Ismail Mulla\email{shaziyaismail.m@gmail.com}
}
\examples{
not yet
}
\seealso{
\code{\link[SOMMD]{pdbaln}}
}
