% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.struct.R
\name{print.struct}
\alias{print.struct}
\title{print.struct}
\usage{
\method{print}{struct}(x, ...)
}
\arguments{
\item{x}{trj object}

\item{...}{additional arguments to be passed to further methods}
}
\value{
Called for its effect.
}
\description{
A short description...
}
\examples{
# Read structure file
struct <- read.struct(system.file("extdata", "HIF2a.gro", package = "SOMMD"))
#Print basic information
print(struct)
}
\author{
Stefano Motta \email{stefano.motta@unimib.it}
}
