% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.trj.R
\name{print.trj}
\alias{print.trj}
\title{TBC}
\usage{
\method{print}{trj}(x, ...)
}
\arguments{
\item{x}{trj object}

\item{...}{additional arguments to be passed to further methods}
}
\description{
A short description...
}
\author{
Stefano Motta \email{stefano.motta@unimib.it}
}
