% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdm.R
\name{rdm}
\alias{rdm}
\title{Reduced size of distance matrix}
\usage{
rdm(pdb, cut.off = 10, inds = NULL, grp = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{pdb}{: a pdb structure object as returned by read.pdb or a numeric vector of ‘xyz’ coordinates.}

\item{cut.off}{: is cut.off radius and it is 10 Angstrom as default}

\item{inds}{: atom and xyz coordinate indices obtained from atom.select that selects the elements of pdb upon which the calculation should be based.}

\item{grp}{: logical, if TRUE atomic distances will be grouped according to their residue membership. See ‘grpby’}

\item{verbose}{:logical, if TRUE possible warnings are printed.}

\item{...}{:arguments passed to and from functions.}
}
\description{
this function only used calculated distance of between atoms which are
closer than cut.off radius to each other in reference frame
}
\author{
Hamid davoudkhani \email{h.davoudkhani@gmail.com}
}
