% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.gro.R
\name{read.gro}
\alias{read.gro}
\title{Function to read gro files}
\usage{
read.gro(file)
}
\arguments{
\item{file}{contains the name and the path to the gro file to be read}
}
\value{
gro the gro file.
}
\description{
Function to read gro files
}
\examples{
#TODO Not added yet
}
\author{
Stefano Motta\email{stefano.motta@unimib.it}
}
