% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dcd.R
\name{read_dcd}
\alias{read_dcd}
\title{Read_dcd: Read CHARMM/X-PLOR/NAMD Binary DCD files}
\usage{
read_dcd(trjfile, big = FALSE, verbose = TRUE, cell = FALSE)
}
\arguments{
\item{trjfile}{name of trajectory file to read. A vector if treat a batch of files}

\item{big}{logical, if TRUE attempt to read large files into a big.matrix object}

\item{verbose}{logical, if TRUE print details of the reading process.}

\item{cell}{logical, if TRUE return cell information only. Otherwise, return coordinates.}
}
\value{

}
\description{
This is a wrap around \link[bio3d]{read.dcd} to standardized functions format
commonly used within this package and ease access to  function reading dcd
formatted file.

the definition of argument duplicated from \link[bio3d]{read.dcd}:
}
\examples{
(traj <- read_dcd("./data/REP_001.dcd"))

}
\references{
\url{https://www.debian.org/doc/debian-policy/ch-controlfields.html}.

Note that R does not require encoding in UTF-8, which is a recent
Debian requirement. Nor does it use the Debian-specific sub-format
which allows comment lines starting with #.
}
\seealso{
\code{\link[SOMMD]{read_ncdf}, \link[SOMMD]{write_dcd},
\link[SOMMD]{read_pdb}, \link[SOMMD]{write_ncdf}}
}
\author{
Hamid Davoukhani \email{h.davoudkhani@gmail.com}
}
