% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ncdf.R
\name{read_ncdf}
\alias{read_ncdf}
\title{Read AMBER Binary netCDF files}
\usage{
read_ncdf(
  trjfile,
  headonly = FALSE,
  verbose = TRUE,
  time = FALSE,
  first = NULL,
  last = NULL,
  stride = 1,
  cell = FALSE,
  at.sel = NULL
)
}
\arguments{
\item{trjfile}{name of trajectory file to read. A vector if treat a batch of files}

\item{headonly}{logical, if TRUE only trajectory header information is returned. If FALSE only trajectory coordinate data is return}

\item{verbose}{logical, if TRUE print details of the reading process.}

\item{time}{logical, if TRUE the first and last have the time unit ps; Otherwise the unit is the frame number.}

\item{first}{starting time or frame number to read; If NULL, start from the begining of the file(s).}

\item{last}{read data until last time or frame number; If NULL or equal to -1, read until the end of the file(s).}

\item{stride}{take at every stride frame(s)}

\item{cell}{logical, if TRUE and headonly is FALSE return cell information only. Otherwise, return header or coordinates.}

\item{at.sel}{an object of class ‘select’ indicating a subset of atomic coordinates to be read.}
}
\value{

}
\description{
This is a wrap around \link[bio3d]{read.ncdf} to standardized functions format
commonly used within this package and ease access to  function reading netCDF
formatted file.

the definition of argument duplicated from \link[bio3d]{read.ncdf}:
}
\examples{
nctrj <- read_ncdf("./data/TEST_001.nc")
}
\references{
Grant, B.J. et al. (2006) Bioinformatics 22, 2695–2696.
\url{https://www.unidata.ucar.edu/software/netcdf/ }
\url{https://cirrus.ucsd.edu/~pierce/ncdf/ }
\url{https://ambermd.org/FileFormats.php#netcdf}
}
\seealso{
\code{\link[SOMMD]{read_pdb}, \link[SOMMD]{write_ncdf}}
}
\author{
Hamid Davoukhani \email{h.davoudkhani@gmail.com}
}
