% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_replica.r
\name{read_replicas}
\alias{read_replicas}
\title{Function to read a set of xtc file, and concatenate them.}
\usage{
read_replicas(FILES)
}
\arguments{
\item{FILES}{the xtc files to be read}
}
\value{
TRJ a list, with $coords being the coordinates of the simulation and $starts being the starting frame of each replica (usefull to have it for some analysis)
}
\description{
Function to read a set of xtc file, and concatenate them.
}
\examples{
#TODO Not added yet
}
\author{
Stefano Motta\email{stefano.motta@unimib.it}
}
