% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rio_xtc.R
\name{rio_read_xtc_natoms}
\alias{rio_read_xtc_natoms}
\title{Read xtc trajectory file}
\usage{
rio_read_xtc_natoms(xtc_filename)
}
\arguments{
\item{xtc_filename}{contains the name and the path to the xtc file}
}
\value{
Returns number of atoms in the structure
}
\description{
Function to read a xtc trajectory file
}
\author{
Alessandro Pandini
}
