% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rio_xtc.R
\name{rio_read_xtc_nframes}
\alias{rio_read_xtc_nframes}
\title{Read xtc trajectory file}
\usage{
rio_read_xtc_nframes(xtc_filename)
}
\arguments{
\item{xtc_filename}{contains the name and the path to the xtc file}
}
\value{
Returns number of frames in the trajectory
}
\description{
Function to read an xtc trajectory file
}
\author{
Alessandro Pandini
}
