% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rio_xtc.R
\name{rio_write_xtc}
\alias{rio_write_xtc}
\title{Write xtc trajectory file}
\usage{
rio_write_xtc(xtc_filename, trj)
}
\arguments{
\item{xtc_filename}{contains the name and the path to the xtc file to write}

\item{trj}{trajectory object to save}
}
\value{
Returns status of write execution
}
\description{
Function to write an xtc trajectory file
}
\author{
Alessandro Pandini
}
