% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network.functions.r
\name{rm.net.diag}
\alias{rm.net.diag}
\title{Remove diagonal elements}
\usage{
rm.net.diag(d)
}
\arguments{
\item{d}{the network in the form of a dataframe}
}
\value{
The network with no diagonal
}
\description{
Function to remove the diagonal from a network
}
\author{
Stefano Motta \email{stefano.motta@unimib.it}
}
