% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_matrix.R
\name{scaler}
\alias{scaler}
\title{Scaling the Distance or similar matrices by standardization or normalization}
\usage{
scaler(mat, method, by = "feature")
}
\arguments{
\item{mat}{a given  numerical matrix}

\item{method}{accepts a character string determines what method implemented
\emph{minmax} = transform the range to 0 and
"robust" = ?
"maxabs" = transform to a range of \link{-1 , 1},
"standard" = normalize the distribution with zero-mean and unit-variance}

\item{by}{\emph{feature} scale each feature of matrices independently
\emph{trajaxis} scale each axis of 3d space independently
\emph{whole} applying scaling method to the whole matrix}
}
\value{
a normalized matrices with values between 0 and 1
}
\description{
Scaling the Distance or similar matrices by standardization or normalization
}
\examples{
not provided here yet
}
