% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som.add.circles.r
\name{SOM.add.circles}
\alias{SOM.add.circles}
\title{Function to plot the silhouette profile for the Nclus cluster of the SOM neurons}
\usage{
SOM.add.circles(SOM, property, scale = 1)
}
\arguments{
\item{SOM:}{the SOM object}

\item{property:}{a vector containing the per-neuron property to plot}

\item{scale:}{a number to scale up or down the size of the drawn circles}
}
\description{
Function to plot the silhouette profile for the Nclus cluster of the SOM neurons
}
\author{
Stefano Motta\email{stefano.motta@unimib.it}
}
