% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som.add.clusters.legend.r
\name{som.add.clusters.legend}
\alias{som.add.clusters.legend}
\title{Add legend clusters}
\usage{
som.add.clusters.legend(N.clus, color.scale)
}
\arguments{
\item{N.clus}{the number of clusters to which put the legent}

\item{color.scale}{the color scale used for the image}
}
\description{
Function to apply a legend of clusters to a SOM map image
}
\examples{
som.add.clusters.legend(N.clus=8, color.scale=COL.SCALE)
}
\author{
Stefano Motta \email{stefano.motta@unimib.it}
}
