% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som.R
\name{somgrid}
\alias{somgrid}
\title{wrapping the \code{\link[kohonen]{somgrid}}function to avoid any errors}
\usage{
somgrid(
  xdim = 8,
  ydim = 6,
  topo = c("rectangular", "hexagonal"),
  neighbourhood.fct = c("bubble", "gaussian"),
  toroidal = FALSE
)
}
\arguments{
\item{xdim}{dimensions of the grid, horizontal axis.}

\item{ydim}{dimensions of the grid, vertical axis.}

\item{topo}{choose between a hexagonal or rectangular topology.}

\item{neighbourhood.fct}{choose between bubble and gaussian neighbourhoods when training a SOM.}

\item{toroidal}{logical, whether the grid is toroidal or not.
If not provided to the unit.distances function,
the information in the grid object will be used.}

\item{grid}{an object of class \code{somgrid}.}
}
\value{

}
\description{
wrapping the \code{\link[kohonen]{somgrid}}function to avoid any errors
}
