% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stride.trj.r
\name{stride.trj}
\alias{stride.trj}
\title{Stride a trj}
\usage{
stride.trj(traj, stride)
}
\arguments{
\item{traj}{a trj object.}

\item{stride}{the stride to apply to the trajectory}
}
\value{
the trj object with a frame every \code{stride}
}
\description{
Apply a stride to the frame of a trj object to reduce the number of frames
}
\examples{
# keep a frame every 10 frame
traj <- stride.trj(traj, 10)

}
\author{
Stefano Motta \email{stefano.motta@unimib.it}
}
