% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.struct.r
\name{summary.struct}
\alias{summary.struct}
\title{Summarizing a structure object}
\usage{
\method{summary}{struct}(object, ...)
}
\arguments{
\item{object}{struct object}

\item{...}{additional arguments to be passed to further methods}
}
\description{
summary method for class struct
}
\author{
Stefano Motta \email{stefano.motta@unimib.it}
}
