% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.trj.R
\name{summary.trj}
\alias{summary.trj}
\title{Summarizing a trajectory object}
\usage{
\method{summary}{trj}(object)
}
\description{
‘summary’ method for class ‘"trj"’.
}
\author{
Stefano Motta \email{stefano.motta@unimib.it}
}
