% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trj2pdb.r
\name{trj2pdb}
\alias{trj2pdb}
\title{Extract frame to pdb}
\usage{
trj2pdb(traj, frame, filename)
}
\arguments{
\item{traj}{a trj object.}

\item{frame}{the frame to extract.}

\item{filename}{for the output pdb file}
}
\value{
a pdb object of the selected frame
}
\description{
Extract a trj frame to a pdb object
}
\examples{
trj2pdb(traj = trj, frame=1000, filename = "Frame_1000.pdb")

#Estract the representative conformation of Neuron 8
NEUR_repres <- neur.representatives(SOM)
trj2pdb(traj = trj, frame=NEUR_repres[8], filename = "../output/Neuron_8.pdb")

}
\author{
Stefano Motta \email{stefano.motta@unimib.it}
}
