% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.trj.R
\name{trj2xyz}
\alias{trj2xyz}
\title{Convert Trajectory to xyz}
\usage{
trj2xyz(trj, inds = NULL)
}
\arguments{
\item{trj}{an object with class trj}

\item{inds}{indices for the output coordinates}
}
\value{
a xyz matrix with frames on rows and coordinates as columns
}
\description{
Convert the trj coordinates 3D-array in a 2D matrix.
}
\examples{
#Read trajectory
trj <- read.trj(trjfile = system.file("extdata", "HIF2a-MD.xtc", package = "SOMMD"),
  topfile = system.file("extdata", "HIF2a.gro", package = "SOMMD"))
trj2xyz(trj)

}
\author{
Stefano Motta \email{stefano.motta@unimib.it}
}
