% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som.add.clusters.legend.r
\name{TwoLetters}
\alias{TwoLetters}
\title{Function to plot the silhouette profile for the Nclus cluster of the SOM neurons}
\usage{
TwoLetters(LETTERS = LETTERS)
}
\arguments{
\item{SOM:}{the SOM object}

\item{scale:}{a number to scale up or down the size of the text}

\item{col:}{the color of the text}
}
\description{
Function to plot the silhouette profile for the Nclus cluster of the SOM neurons
}
\author{
Stefano Motta\email{stefano.motta@unimib.it}
}
