% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som.R
\name{unit.distances}
\alias{unit.distances}
\title{a wrapper over \link[kohenen]{unit.distances}}
\usage{
unit.distances(grid, torodial)
}
\arguments{
\item{grid}{an object of class \code{somgrid}.}

\item{torodial}{logical, whether the grid is toroidal or not.
If not provided to the unit.distances function,
the information in the grid object will be used.}
}
\value{
Function unit.distances returns a (symmetrical) matrix containing distances.
When grid$n.hood equals "circular", Euclidean distances are used; for
grid$n.hood is "square" maximum distances. For toroidal maps
(joined at the edges) distances are calculated for the shortest path.
}
\description{
a wrapper over \link[kohenen]{unit.distances}
}
