% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_utils.R
\name{add_evolution_trace}
\alias{add_evolution_trace}
\title{Title}
\usage{
add_evolution_trace(som_obj, path, line_thickness = 3, point_size = 2)
}
\arguments{
\item{som_obj}{should be a \code{kohonen} object}

\item{path}{the path of neuron want to show on the plot}

\item{line_thickness}{the thickness of connecting lines same as \code{lwd} argument in \code{lines}}

\item{point_size}{the size of points same as  arguments \code{cex} argument in \code{points}}
}
\value{

}
\description{
Title
}
\author{
Hamid Davoukhani \email{h.davoudkhani@gmail.com}
}
