% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_utils.R
\name{add_neuron_population}
\alias{add_neuron_population}
\title{add populaiton of each neuron to the som plots}
\usage{
add_neuron_population(som_obj, text = FALSE, shape = c("hex", "round"))
}
\arguments{
\item{som_obj}{should be a \code{kohonen} object}

\item{text}{logical , if TRUE print the population of each neuron as number}

\item{shape}{a chosen shape size of which show the popluation of each neurons}
}
\value{

}
\description{
add populaiton of each neuron to the som plots
}
\author{
Hamid Davoukhani \email{h.davoudkhani@gmail.com}
}
