% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_utils.R
\name{add_neuron_population}
\alias{add_neuron_population}
\title{add populaiton of each neuron to the som plots}
\usage{
add_neuron_population(som_obj, text = FALSE, shape = c("hex", "round"))
}
\arguments{
\item{som_obj}{should be a \code{kohonen} object}

\item{text}{logical , if TRUE print the population of each neuron as number}

\item{shape}{a chosen shape size of which show the popluation of each neurons}
}
\value{

}
\description{
add populaiton of each neuron to the som plots
}
\examples{
###
library(kohonen)
data(wines)
som_wines <- som(scale(wines), grid = somgrid( 6 , 6, "hexagonal"))
plot_clustered_map (som_wines,
cluster_method = "hier",
clustering_parameter = c( 5 , "ward.D"  ),
shape = "straight")
add_neuron_population(som_wines , shape = "round" )
}
\author{
Hamid Davoukhani \email{h.davoudkhani@gmail.com}
}
