% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average.neur.property.r
\name{average.neur.property}
\alias{average.neur.property}
\title{Compute average property}
\usage{
average.neur.property(SOM, P)
}
\arguments{
\item{SOM:}{the SOM object to cluster}

\item{P:}{the property for each frame of the simulation}
}
\value{
avg.neur.p the per-neuron average of the property.
}
\description{
Function to compute the average value of a property for each neuron.
}
\examples{
#Compute distances between these two atoms in every frame of the simulation
Term_dist <- apply(trj$coord[Terminals,,], 3, dist)
#Compute average property value for each neuron
average.neur.property(SOM, Term_dist)

}
\author{
Stefano Motta \email{stefano.motta@unimib.it}
}
