% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_neurons.R
\name{cluster_neuron}
\alias{cluster_neuron}
\title{Find optimal clusters.}
\usage{
cluster_neuron(
  som_obj,
  min_cluster = 6,
  max_cluster = 15,
  kohonen_object = NULL
)
}
\arguments{
\item{som_obj}{should be a \code{kohonen} object, a \code{matrix}, or
\code{dist} object of \link[stats]{dist()}}

\item{min_cluster}{minimum number of clusters shoulde be evaluate}

\item{max_cluster}{upper bond of number of clusters.}

\item{kohonen_object}{if the \code{som-obj} was not as a \code{kohonen}
object it can call here to plot clustered som map}
}
\value{
a list consist of all clusters content such as neuron clusters number, silhouette score and its summary.
}
\description{
evaluate the number of  different clusters with indexes such as silhouette,
it also plot for each clusters
}
\examples{
library(kohonen)
data("wines")
som_wines <- som(scale(wines), grid = somgrid(5, 5, "hexagonal"))
(clusters_result <- cluster_neuron(som_wines, 3 , 7)
}
\author{
Hamid davoudkhani \email{h.davoudkhani@gmail.com}
}
