% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_utils.R
\name{highlight_a_cluster}
\alias{highlight_a_cluster}
\title{Highlight a cluster and its attributes}
\usage{
highlight_a_cluster(
  som_obj,
  clusters,
  cluster_number,
  label = NULL,
  property_value = NULL,
  property_color = NULL,
  label_color = "black",
  col = "red",
  lwd = 4,
  ...
)
}
\arguments{
\item{som_obj}{should be a \code{kohonen} object}

\item{clusters}{a vector of cluster neuron}

\item{cluster_number}{the cluster to be highlighted}

\item{label}{label print at the center of cluster}

\item{property_value}{a different value print at center of different neuron of cluster}

\item{property_color}{color property note, if \code{NULL} use \code{label_color}}

\item{label_color}{color of label note, default is \code{black}}

\item{col}{the color draw a boundary around the cluster}

\item{lwd}{thickness of both notes}

\item{...}{}
}
\value{

}
\description{
Highlight a clusters and add note at center of cluster and a property value in a different center of clusters
}
\examples{
library(kohonen)
data(wines)
som_wines <- som(scale(wines), grid = somgrid( 6 , 6, "hexagonal"))
plot_clustered_map (som_wines,
 cluster_method = "hier",
 clustering_parameter = c( 5 , "ward.D"  ),
 shape = "straight" )
 highlight_cluster(som.wines)
}
