% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/native_contacts.r
\name{native_contacts}
\alias{native_contacts}
\title{Function to select only distances between residues making contacts in reference file or a frame of the simulation}
\usage{
native_contacts(coord, distance, MOL2 = FALSE, atoms = NULL)
}
\arguments{
\item{coord}{matrix of N atomic coordinates (N rows, 3 columns) that will be used to determine the native_contacts}

\item{distance}{the distance cut-off}

\item{MOL2}{can be FALSE (default), use the whole distance matrix, or a vector containing the atomic number of the second molecule (and compute only intermolecular disatnces)}

\item{atoms}{can be FALSE (default), consider all the atoms present in coords, or a vector containing a set of atomic numbers to consider in the calculation (e.g. only CB). atoms can be obtained with the bio3d atom.select function}
}
\value{
SELE the selection of distances
}
\description{
Function to select only distances between residues making contacts in reference file or a frame of the simulation
}
\author{
Stefano Motta\email{stefano.motta@unimib.it}
}
