% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.silhouette.score.r
\name{plot.silhouette.score}
\alias{plot.silhouette.score}
\title{Function to plot the silhouette score for the clustering of SOM neurons}
\usage{
\method{plot}{silhouette.score}(
  SOM,
  dist_clust = "euclidean",
  clust_method = "complete",
  intervall = seq(2, 30)
)
}
\arguments{
\item{SOM}{the SOM object to cluster}

\item{dist_clust}{the metric for the distance calculation}

\item{clust_method}{the method for the clustering (passed to the hclust function}

\item{intervall}{the cluster number on which the silhouette score will be computed}
}
\value{
D the set of distances used to train the SOM is computed for all the frames.
}
\description{
Function to plot the silhouette score for the clustering of SOM neurons
}
\author{
Stefano Motta\email{stefano.motta@unimib.it}
}
