% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_property.R
\name{plot_property}
\alias{plot_property}
\title{Title}
\usage{
plot_property(
  som_obj,
  property,
  FUN = mean,
  main = NULL,
  heatkey = TRUE,
  palette_name = NULL,
  ...
)
}
\arguments{
\item{som_obj}{a \code{kohonen} object}

\item{property}{a vector of property for either each neurons or each observation}

\item{FUN}{when length of property vector is equal to number of observation,
this given argument use as a function to aggregate values for each neurons}

\item{main}{the title of the plot}

\item{heatkey}{logical, when is true, plot the spectrum of the property}

\item{palette_name}{palette color used to hue neurons}

\item{...}{other graphical arguments can be passed to plot funtion}
}
\value{

}
\description{
Title
}
\examples{
 load("data/pdb_and_tarj_object.RData")
 som_trj <- SOMMD::som(trj1 , grid =  somgrid( 6 , 6, "hexagonal"))
 ##  property vector for each neurons
 dummy_property_neurons <-  rnorm(36 , mean = 8 , sd  = 4)
 plot_property(som_trj , property = dummy_property_nuerons )

 ##  property vector for each instance
 dummy_property_observation <-  rnorm(102 , mean = 8 , sd  = 4)
 plot_property(som_trj , property = dummy_property_observation, FUN = median ,shape = "straight",
 palette_name =  colorRampPalette(c("blue", "white", "red")))
}
