% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreAlignTrajectory.R
\name{Pre_algn_pdbaln}
\alias{Pre_algn_pdbaln}
\title{ReQ 3: Pre Align the Trajectory with pdbaln}
\usage{
Pre_algn_pdbaln(
  files,
  fit = FALSE,
  pqr = FALSE,
  ncore = 1,
  nseg.scale = 1,
  progress = NULL,
  ...
)
}
\arguments{
\item{files}{a character vector of PDB file names. Alternatively, a list of pdb objects can be provided.}

\item{fit}{fit logical, if TRUE coordinate superposition is performed on the input structures.}

\item{pqr}{pqr logical, if TRUE the input structures are assumed to be in PQR format.}

\item{ncore}{ncore number of CPU cores used to do the calculation. ncore>1 requires package ‘parallel’ installed.}

\item{nseg.scale}{nseg.scale split input data into specified number of segments prior to running multiple core calculation. See fit.xyz.}

\item{progress}{progress progress bar for use with shiny web app.}

\item{......}{extra arguments passed to seqaln function.}
}
\value{

}
\description{
This is a wrap around \link[bio3d]{pdbaln}
This Creates multiple sequences alignments from a list of PDB files returning aligned sequence and structure records
This is written for re-aligns the trajectory,  using output of requirement (REQ-2) as input ,
Returns a list of class "pdbs" with the five  components: xyz, resno, b, chain, id, ali, and call.
The coordinates of the aligned trajectory.
the definition of argument duplicated from \link[bio3d]{pdbaln}:

@author Shaziya Ismail Mulla\email{shaziyaismail.m@gmail.com}
}
\examples{
files <- get.pdb(c("4q21","5p21"), URLonly=TRUE)




}
\references{
:https://www.rdocumentation.org/packages/bio3d/versions/2.4-2/topics/pdbaln
https://rdrr.io/cran/bio3d/man/pdbaln.html
https://search.r-project.org/CRAN/refmans/bio3d/html/pdbaln.html
}
\seealso{
\code{\link[SOMMD]{pdbaln}}
}
