% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.trj.R
\name{print.trj}
\alias{print.trj}
\title{Print Trajectory}
\usage{
\method{print}{trj}(x, ...)
}
\arguments{
\item{x}{trj object}

\item{...}{additional arguments to be passed to further methods}
}
\value{
Called for its effect.
}
\description{
A short description...
}
\examples{
#Read trajectory
trj <- read.trj(trjfile = system.file("extdata", "HIF2a-MD.xtc", package = "SOMMD"),
  topfile = system.file("extdata", "HIF2a.gro", package = "SOMMD"))
#Print basic informations
print(trj)
}
\author{
Stefano Motta \email{stefano.motta@unimib.it}
}
