% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.gro.R
\name{read.gro}
\alias{read.gro}
\title{Read gro file}
\usage{
read.gro(file)
}
\arguments{
\item{file}{contains the name and the path to the gro file to be read}
}
\value{
Returns a list of class "gro" with the following components:

\item{atom}{ a data frame containing all atomic coordinate with a row per atom and a column per record type.}

\item{xyz}{ a numeric matrix of class "xyz" containing the atomic coordinate data.}

\item{box}{ a vector of box size.}

\item{call}{ the matched call.}
}
\description{
Function to read gro files
}
\author{
Stefano Motta \email{stefano.motta@unimib.it}
}
