% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som.add.numbers.r
\name{som.add.numbers}
\alias{som.add.numbers}
\title{Add Neuron Numbering}
\usage{
som.add.numbers(SOM, scale = 1, col = "black")
}
\arguments{
\item{SOM:}{the SOM object}

\item{scale:}{a number to scale up or down the size of the text}

\item{col:}{the color of the text}
}
\description{
Add the neuron numbering scheme to the SOM plot
}
\examples{
som.add.numbers(SOM, scale=0.5, col="black")
}
\author{
Stefano Motta \email{stefano.motta@unimib.it}
}
