% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stride.trj.R
\name{stride.trj}
\alias{stride.trj}
\title{Stride a trj}
\usage{
stride.trj(trj, stride)
}
\arguments{
\item{trj}{a trj object.}

\item{stride}{the stride to apply to the trajectory}
}
\value{
An object of class trj with a frame every \code{stride}
}
\description{
Apply a stride to the frame of a trj object to reduce the number of frames
}
\examples{
# Read the simulation
trj <- read.trj(trjfile = system.file("extdata", "HIF2a-MD.xtc", package = "SOMMD"),
  topfile = system.file("extdata", "HIF2a.gro", package = "SOMMD"))
# keep a frame every 2 frame
trj_strd <- stride.trj(trj, 2)

}
\author{
Stefano Motta \email{stefano.motta@unimib.it}
}
