% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct2pdb.R
\name{struct2pdb}
\alias{struct2pdb}
\title{Convert structure to pdb object}
\usage{
struct2pdb(struct)
}
\arguments{
\item{struct}{contains the struct object to convert}
}
\value{
Returns an object with class "pdb"

An object of class "pdb"
}
\description{
Convert a struct object into a pdb obtect
}
\examples{
# Read structure file 
struct <- read.struct(system.file("extdata", "HIF2a.gro", package = "SOMMD"))
#Convert structure to pdb object
pdb <- struct2pdb(struct)

}
\author{
Stefano Motta \email{stefano.motta@unimib.it}
}
